+++ ARCKG/pair.py
@@ -54,8 +54,9 @@ class Pair:
         folder = node_id_to_folder_path(self.node_id, semantic_memory_root)
         os.makedirs(folder, exist_ok=True)
         path = id_to_json_path(self.node_id, semantic_memory_root)
-        with open(path, "w") as f:
-            json.dump({"id": self.node_id, "result": self.to_json()}, f, indent=2)
+        if not os.path.exists(path):
+            with open(path, "w") as f:
+                json.dump({"id": self.node_id, "result": self.to_json()}, f, indent=2)
 
         if self.input_grid is not None:
             self.input_grid.save(semantic_memory_root)
